Change test execution method to run_dag() from operator.execute()
