+++ python-sdk/tests/data/data_validation.csv
@@ -1,4 +1,4 @@
 name,age,city,emp_id
-Dwight Schrute,30,,10
+Dwight Schrute,30.0,,10
 Michael Scott,,LA,1
 Jim Halpert,,California City,35
